updated response codes

+++ api/review/delete.php
@@ -10,7 +10,9 @@ $validator = new Validator;
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
     // Authenticating user  
-    $user = $fun->verify_token();
+    if (!empty($fun)) {
+        $user = $fun->verify_token();
+    }
 
     $request = file_get_contents("php://input");
     $request = json_decode($request);
@@ -32,7 +34,9 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
     // deleting category from database  
     try {
-        $db->from('review')->Where("id")->is($request->id)->delete();
+        if (!empty($db)) {
+            $db->from('review')->Where("id")->is($request->id)->delete();
+        }
         echo json_encode(["status" => true, "msg" => "review Deleted"]);
     } catch (Exception $ex) {
         echo json_encode(["success" => false, "msg" => $ex->getMessage()]);
